% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_output.R
\name{lm_output}
\alias{lm_output}
\title{Extract and format the output of a linear regression object.}
\usage{
lm_output(model, var_index = 2, res_df = model$df.residual, digit = 3,
  mode = NULL, cluster = NULL)
}
\value{
TBA
}
\description{
Extract the coeffients, standard  errors and p-values from a linear regression object (p-values based on t-dist (t-score under H0 - mean:0 & SE) with residual DF (# of clusters - # of coefficients)
}
